students/k33391/Bartalevich_Liza/lab-3/lab/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.8.02
Django==4.2.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-spectacular==0.26.5
ghp-import==2.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Markdown==3.5.1
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
watchdog==3.0.0
